feat: created field component
created field component to use as normal field or password field and changed all html inputs to new
field component
