+++ src/presentation/components/field/styles.sass
@@ -1,0 +1,39 @@
+@use '../../assets/sass/variables'
+@use '../../assets/sass/mixins'
+
+.input
+    border: none
+    border-radius: 10px
+    box-sizing: border-box
+    font-size: 1.2rem
+    margin-bottom: 1rem
+    padding: 0.5rem
+    width: 100%
+
+    &--icon
+        margin-bottom: 0
+        padding: 0.5rem 3.5rem 0.5rem 0.5rem
+
+        &:focus
+            outline: 2px solid variables.$black
+
+.password-container
+    position: relative
+
+    &__button
+        align-items: center
+        background-color: variables.$white
+        border: none
+        bottom: 0
+        display: flex
+        justify-content: center
+        position: absolute
+        right: 1rem
+        top: 0
+
+        &--icon
+            height: variables.$iconHeight
+            width: variables.$iconWidth
+
+        @include mixins.desktop
+            cursor: pointer
