+++ src/presentation/pages/login/styles.sass
@@ -44,18 +44,6 @@
             font-size: 1.2rem
             margin-bottom: 0.5rem
 
-        &__input
-            border: none
-            border-radius: 10px
-            box-sizing: border-box
-            font-size: 1.2rem
-            margin-bottom: 1rem
-            padding: 0.5rem
-            width: 100%
-
-            &:focus
-                outline: 2px solid variables.$black
-
         &__link
             align-self: flex-end
             font-size: 1.2rem
